ISIS-1976: add layer of abstraction as first step to deprecate ObjectAdapter caching
Task-Url: https://issues.apache.org/jira/browse/ISIS-1976
